+++ src/entry/Page.vue
@@ -4,7 +4,7 @@
             <div class="window-header flex h-10 items-center border-b border-solid border-gray-200">
                 <div class="window-header-title flex-grow flex items-center">
                     <div class="pl-2 py-2">
-                        <img src="/logo.svg" class="w-4 t-4" />
+                        <img src="/logo.svg" class="w-4 h-4" />
                     </div>
                     <div class="p-2 flex-grow truncate overflow-hidden text-ellipsis max-w-96">
                         {{ pageTitle }}